+++ libc/sysv/consts/dt.h
@@ -4,14 +4,14 @@
 #if !(__ASSEMBLER__ + __LINKER__ + 0)
 COSMOPOLITAN_C_START_
 
-extern const long DT_UNKNOWN;
-extern const long DT_FIFO;
-extern const long DT_CHR;
-extern const long DT_DIR;
-extern const long DT_BLK;
-extern const long DT_REG;
-extern const long DT_LNK;
-extern const long DT_SOCK;
+extern const uint8_t DT_UNKNOWN;
+extern const uint8_t DT_FIFO;
+extern const uint8_t DT_CHR;
+extern const uint8_t DT_DIR;
+extern const uint8_t DT_BLK;
+extern const uint8_t DT_REG;
+extern const uint8_t DT_LNK;
+extern const uint8_t DT_SOCK;
 
 COSMOPOLITAN_C_END_
 #endif /* !(__ASSEMBLER__ + __LINKER__ + 0) */
