+++ libc/sysv/consts/prot.h
@@ -4,12 +4,12 @@
 #if !(__ASSEMBLER__ + __LINKER__ + 0)
 COSMOPOLITAN_C_START_
 
-extern const long PROT_NONE;
-extern const long PROT_READ;
-extern const long PROT_WRITE;
-extern const long PROT_EXEC;
-extern const long PROT_GROWSDOWN;
-extern const long PROT_GROWSUP;
+extern const int PROT_NONE;
+extern const int PROT_READ;
+extern const int PROT_WRITE;
+extern const int PROT_EXEC;
+extern const int PROT_GROWSDOWN;
+extern const int PROT_GROWSUP;
 
 COSMOPOLITAN_C_END_
 #endif /* !(__ASSEMBLER__ + __LINKER__ + 0) */
