+++ libc/sysv/consts/splice.h
@@ -4,10 +4,10 @@
 #if !(__ASSEMBLER__ + __LINKER__ + 0)
 COSMOPOLITAN_C_START_
 
-extern const long SPLICE_F_GIFT;
-extern const long SPLICE_F_MORE;
-extern const long SPLICE_F_MOVE;
-extern const long SPLICE_F_NONBLOCK;
+extern const unsigned int SPLICE_F_GIFT;
+extern const unsigned int SPLICE_F_MORE;
+extern const unsigned int SPLICE_F_MOVE;
+extern const unsigned int SPLICE_F_NONBLOCK;
 
 COSMOPOLITAN_C_END_
 #endif /* !(__ASSEMBLER__ + __LINKER__ + 0) */
